Fix handling of rich text rendering
- Re-add conversion of image links to galllery
- Use RichText::getEnhancedHtml() for non editable content
- Move gallery methods to RichText and make them private
- Enable user mentions on itil item creation form
