+++ ajax/compareKbRevisions.php
@@ -34,7 +34,7 @@
  * @since 9.1
  */
 
-use Glpi\Toolbox\RichText;
+use Glpi\RichText\RichText;
 
 include ('../inc/includes.php');
 header("Content-Type: application/json; charset=UTF-8");
