+++ inc/change.class.php
@@ -32,7 +32,7 @@
 
 use Glpi\Application\View\TemplateRenderer;
 use Glpi\ContentTemplates\Parameters\ChangeParameters;
-use Glpi\Toolbox\RichText;
+use Glpi\RichText\RichText;
 
 if (!defined('GLPI_ROOT')) {
    die("Sorry. You can't access this file directly");
@@ -1447,7 +1447,7 @@ class Change extends CommonITILObject {
                $link .= "'>";
                $link .= "<span class='b'>".$change->fields["name"]."</span></a>";
                $link = sprintf(__('%1$s %2$s'), $link,
-                  Html::showToolTip(RichText::getSafeHtml($change->fields['content']),
+                  Html::showToolTip(RichText::getEnhancedHtml($change->fields['content']),
                      ['applyto' => 'change'.$change->fields["id"].$rand,
                         'display' => false]));
