+++ inc/dashboard/widget.class.php
@@ -32,7 +32,7 @@
 
 namespace Glpi\Dashboard;
 
-use Glpi\Toolbox\RichText;
+use Glpi\RichText\RichText;
 use Html;
 use Mexitek\PHPColors\Color;
 use Michelf\MarkdownExtra;
@@ -1824,7 +1824,7 @@ HTML;
 
          $content_size = strlen($entry['content']);
          $content = strlen($entry['content'])
-            ? RichText::getSafeHtml($entry['content']).
+            ? RichText::getEnhancedHtml($entry['content']).
               ($content_size > 300
                ? "<p class='read_more'><span class='read_more_button'>...</span></p>"
                : ""
