+++ inc/gantt/datafactory.class.php
@@ -32,7 +32,7 @@
 
 namespace Glpi\Gantt;
 
-use Glpi\Toolbox\RichText;
+use Glpi\RichText\RichText;
 
 if (!defined('GLPI_ROOT')) {
    die("Sorry. You can't access this file directly");
