+++ templates/components/itilobject/timeline/form_followup.html.twig
@@ -40,7 +40,10 @@
 {% block timeline_card %}
    {% if form_mode == 'view' %}
       <div class="read-only-content">
-         {{ entry_i['content']|safe_html }}
+         {{ entry_i['content']|enhanced_html({
+            'user_mentions': true,
+            'images_gallery': true
+         }) }}
 
          <div class="timeline-badges">
             {% if entry_i['sourceitems_id'] %}
