+++ templates/components/itilobject/timeline/form_task.html.twig
@@ -48,7 +48,10 @@
 
          <div class="text-content" data-bs-html="true" data-bs-custom-class="todo-list-tooltip"
               title="{{ entry_i['content']|html_to_text }}" data-bs-toggle="tooltip">
-            {{ entry_i['content']|safe_html }}
+            {{ entry_i['content']|enhanced_html({
+               'user_mentions': true,
+               'images_gallery': true
+            }) }}
          </div>
 
          <div class="timeline-badges">
