+++ templates/components/itilobject/timeline/main_description.html.twig
@@ -77,7 +77,10 @@
                      </div>
                   {% endif %}
                   {% if not itiltemplate.isHiddenField('content') %}
-                     {{ item.fields['content']|safe_html }}
+                     {{ item.fields['content']|enhanced_html({
+                        'user_mentions': true,
+                        'images_gallery': true
+                     }) }}
                   {% endif %}
                </div>
                <div class="edit-content collapse">
