+++ templates/components/itilobject/timeline/simple_form.html.twig
@@ -68,6 +68,7 @@
          field_options|merge({
             'enable_richtext': true,
             'enable_fileupload': true,
+            'enable_mentions': true,
             'uploads': uploads,
             'add_field_class': 'col-12 itil-textarea-content',
          })
